"TAGS" implemented. Compact collector Form UI for Tags Added
The collector form now supports the use of "Tags" [Star | Square | Triangle | Circle] these can be used as custom markers for whatever the user wants.

The setting form also has a setting to give tags custom names. These names are saved in the settings datafile.

The information of cards "tagged" will be save in its own dedicated save file.
A new "Save Files" folder has been added in the root folder of the app to store both the main Save File and the Tags Save File.