Collector: Music now plays from a playlist
The collector form now has a "Playlist" of songs to play. This list is shuffle the first time a collector form is open and plays the first song on the list. From there, once a song finish playing it will play the next song on the list until the playlist is completed. Once that happens the playlist shuffles again and the whole process repeats.

A new song added: "LibraryMenu"