New Launcher Form Implemented  + Year DB Selection + More User Friendly Loads for Stats/Pricing Reports
New Launcher form will allow to separate the Load DB control from the main menu.

Launcher Form has now the option to select a "Year DB" to load a DB with only the setcards released BY the selected years. Stats will also reflect this year limit.

Added a Shown event to the Stats and Pricing reports to show the Form (empty) while the components are loaded with a "Loading..." mensage on top. This way the user is not sitting there with no visible screen while these reports load.