Collector Forms: Setpack list will not show empty sets.
Sets in the DB that have no cards will not be displayed in the setpack list component.
if a Set Group has no Sets (this can happen when loading an older DB year) the set list will have "NO Sets in this group." as the only item in the list. Open Set and Open Set $$ will not work while "NO Sets in this group." is selected.