Deck Builder: Open Pricing Report Button only available if Main Deck is not empty
This restriction will prevent the user from entering the Pricing Report if the Deck contains no Main Deck Cards. This is to prevent the Pricing Report Issues with the UI, also, is seems very unnecesary to check the pricing report of a empty deck even if it has Extra/Side deck cards.