Bug Fix: Fail to update the Owned status of a Mastercard when marking removing the "Owned" marks of the only owned setcard.
When having a Mastercard with a Single Setcard owned, and then you remove that one Setcard from Owned status, the application failed to removed the Owned status FROM THE MASTERCARD OBJECT making it seem like we still own that card. This was just a logic oversight and it has been fixed.